requirements.txt
aiohttp==3.10.11
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
djangorestframework==3.14.0
websockets==13.1
